add dark mode handling to network_followers plot
